+++ lib/payment_server_web/graphql/types/sent_money_type.ex
@@ -5,7 +5,7 @@ defmodule PaymentServerWeb.Graphql.Types.SentMoneyType do
   @desc "Send money object type"
   object :sent_money do
     field :sender_wallet_id, :id
-    field :amount_received, :float
+    field :amount_received, :decimal
     field :wallet, :wallet
   end
 end
